shopglobal_text_search/requirements.txt
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
black==23.11.0
flake8==6.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
gunicorn==21.2.0
google-cloud-discoveryengine